+++ src/main/java/com/global/api/entities/enums/TaxType.java
@@ -2,8 +2,11 @@ package com.global.api.entities.enums;
 
 public enum TaxType implements IStringConstant {
     NotUsed("NOTUSED"),
-    SalesTax("SALES_TAX"),
-    TaxExempt("TAXEXEMPT");
+    SalesTax("SALESTAX"),
+    TaxExempt("TAXEXEMPT"),
+    Not_Used("NOT_USED"),
+    Sales_Tax("SALES_TAX"),
+    Tax_Exempt("TAX_EXEMPT");
 
     String value;
     TaxType(String value) {